requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# CrewAI and AI
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# NLP and AI
spacy

# Document processing
unstructured
python-docx
openpyxl
pypdf

# Additional dependencies for document processing
python-pptx

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0

# CORS support
python-multipart
